changed radius from int32 to float

+++ Engine Architecture Assessment/Circle.cpp
@@ -38,7 +38,7 @@ void Circle::Draw()
 	
 }
 
-CircleParameters::CircleParameters(int32_t radius)
+CircleParameters::CircleParameters(float radius)
 {
 	this->radius = radius;
 }
